I tried but failed miserably
I studied a while book about /proc file system later to realize that it
was for kernel 2.6 now I dont know where to get started from for
specifically learning about linux driver. The book in the readme is not
detailed and I cannot understand even the inch.

So it most prob ends with I tried!